+++ src/core/server/views/dealership.ts
@@ -5,6 +5,8 @@ import { View_Events_Dealership } from '../../shared/enums/views';
 import { BlipController } from '../systems/blip';
 import { HologramController } from '../systems/hologram';
 import { VehicleData } from '../../shared/information/vehicles';
+import { playerFuncs } from '../extensions/Player';
+import { vehicleFuncs } from '../extensions/Vehicle';
 
 class DealershipFunctions {
     static init() {
@@ -37,63 +39,63 @@ class DealershipFunctions {
                 model: dealership.vehiclePreview
             });
 
-            BlipController.add({
-                text: `${dealership.class} Dealership`,
-                color: 13,
-                sprite: 225,
-                scale: 1,
-                shortRange: true,
-                pos: dealership.position,
-                uid: `dealership-${i}`
-            });
+            if (dealership.createBlip) {
+                BlipController.add({
+                    text: `${dealership.name}`,
+                    color: 13,
+                    sprite: 225,
+                    scale: 1,
+                    shortRange: true,
+                    pos: dealership.position,
+                    uid: `dealership-${i}`
+                });
+            }
         }
+    }
+
+    static purchase(player: alt.Player, model: string) {
+        if (!player || !player.valid) {
+            return;
+        }
+
+        if (!model) {
+            playerFuncs.emit.soundFrontend(player, 'Hack_Failed', 'DLC_HEIST_BIOLAB_PREP_HACKING_SOUNDS');
+            return;
+        }
+
+        const vehicleData = VehicleData.find((x) => x.name === model);
+        if (!vehicleData || !vehicleData.sell) {
+            playerFuncs.emit.soundFrontend(player, 'Hack_Failed', 'DLC_HEIST_BIOLAB_PREP_HACKING_SOUNDS');
+            return;
+        }
+
+        if (player.data.bank + player.data.cash < vehicleData.price) {
+            playerFuncs.emit.notification(player, `~r~Invalid Balance`);
+            playerFuncs.emit.soundFrontend(player, 'Hack_Failed', 'DLC_HEIST_BIOLAB_PREP_HACKING_SOUNDS');
+            return;
+        }
+
+        if (!playerFuncs.currency.subAllCurrencies(player, vehicleData.price)) {
+            playerFuncs.emit.notification(player, `~r~Invalid Balance`);
+            return;
+        }
+
+        vehicleFuncs.new.add(
+            player,
+            {
+                model: vehicleData.name,
+                fuel: 100,
+                position: { x: 0, y: 0, z: 0 },
+                rotation: { x: 0, y: 0, z: 0 },
+                color: { r: 255, g: 255, b: 255, a: 255 }
+            },
+            false
+        );
+
+        playerFuncs.emit.notification(player, `Visit a garage to spawn your new vehicle.`);
+        playerFuncs.emit.sound2D(player, 'item_purchase');
     }
 }
 
+alt.onClient(View_Events_Dealership.Purchase, DealershipFunctions.purchase);
 DealershipFunctions.init();
-
-// function handlePurchase(player: alt.Player, model: string, color: alt.RGBA) {
-//     if (!player || !player.valid) {
-//         return;
-//     }
-
-//     if (!model) {
-//         return;
-//     }
-
-//     const vehicleData = VehicleData.find((veh) => veh.name === model);
-//     if (!vehicleData) {
-//         return;
-//     }
-
-//     if (!vehicleData.sell) {
-//         playerFuncs.emit.message(player, `Nice try... but the ${model} is not for sale.`);
-//         return;
-//     }
-
-//     if (!DEFAULT_CONFIG.VEHICLE_DEALERSHIP_SPAWNS[vehicleData.class]) {
-//         playerFuncs.emit.message(player, `Notify Admin that vehicle class... ${vehicleData.class} has no spawn point.`);
-//         return;
-//     }
-
-//     if (!playerFuncs.currency.subAllCurrencies(player, vehicleData.price)) {
-//         playerFuncs.emit.message(player, `Not enough money for this vehicle. \$${vehicleData.price}.`);
-//         return;
-//     }
-
-//     vehicleFuncs.new.add(player, {
-//         model: vehicleData.name,
-//         fuel: 100,
-//         position: DEFAULT_CONFIG.VEHICLE_DEALERSHIP_SPAWNS[vehicleData.class],
-//         rotation: { x: 0, y: 0, z: 0 },
-//         color
-//     });
-
-//     playerFuncs.emit.message(
-//         player,
-//         `Vehicle has been added to your personal vehicles. Check your phone to locate it.`
-//     );
-
-//     playerFuncs.emit.notification(player, `~r~-\$${vehicleData.price}`);
-//     playerFuncs.emit.sound2D(player, 'item_purchase');
-// }